+++ src/core/StelViewportEffect.hpp
@@ -42,7 +42,7 @@ public:
 	virtual void paintViewportBuffer(const QOpenGLFramebufferObject* buf) const;
 	//! Distort an x,y position according to the distortion.
 	//! The default implementation does nothing.
-	virtual void distortXY(float& x, float& y) const {Q_UNUSED(x); Q_UNUSED(y);}
+	virtual void distortXY(qreal& x, qreal& y) const {Q_UNUSED(x); Q_UNUSED(y);}
 };
 
 
@@ -53,7 +53,7 @@ public:
 	~StelViewportDistorterFisheyeToSphericMirror();
 	virtual QString getName() const {return "sphericMirrorDistorter";}
 	virtual void paintViewportBuffer(const QOpenGLFramebufferObject* buf) const;
-	virtual void distortXY(float& x, float& y) const;
+	virtual void distortXY(qreal& x, qreal& y) const;
 private:
 	const int screen_w;
 	const int screen_h;
@@ -63,7 +63,7 @@ private:
 
 	Vec2f *texture_point_array;
 	int max_x,max_y;
-	double step_x,step_y;
+	qreal step_x,step_y;
 
 	QVector<Vec2f> displayVertexList;
 	QVector<Vec4f> displayColorList;
